fiscia-pro/requirements.txt
click==8.1.7
pydantic==2.7.1
fastapi==0.110.2
uvicorn[standard]==0.30.1
httpx==0.27.0
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1
PyJWT==2.8.0
bcrypt==4.1.3
aiosqlite==0.20.0
